20200621 交易日历可以导入到腾讯云

+++ Test/stock_daily_to_db.py
@@ -14,35 +14,44 @@ show_func = print if show else lambda a: a
 
 if __name__ == "__main__":
     starttime = datetime.datetime.now()
-    Chinadaily = Getdailyfromtscode('','20190101','20200201')
-    Chinadaily['trade_date'] = pd.to_datetime(Chinadaily['trade_date'],format='%Y%m%d')
-    # show_func(Chinadaily)
-    # Tocsv(Chinadaily, '', '1days')
+    # 首先获取一下数据库连接
     engine = connect_db_engine()
-    try:
-        Chinadaily.to_sql('stock_china_daily_temp', con=engine, if_exists='replace', index=False)
-    except Exception as e:
-        print(e)
-    print('数据已经导入临时表，共导入数据%s条' % (str(Chinadaily.shape[0])))
-    try:
-        query_sql = """
-                      INSERT INTO `stock_china_daily` (`ts_code`, `trade_date`, `open`, `high`, `low`, `close`, `pre_close`, `change`, `pct_chg`, `vol`, `amount`)
-            select
-            	`ts_code`, `trade_date`, `open`, `high`, `low`, `close`, `pre_close`, `change`, `pct_chg`, `vol`, `amount` from stock_china_daily_temp;
-                      """
-        query_sql2 = """
-                delete from stock_china_daily_temp
-            """
-        engine.execute(query_sql)
-        print('数据已经导入到正是表')
-        engine.execute(query_sql2)
-        print('临时表数据已经删除')
-    except Exception as ae:
-        print(ae)
-    engine.dispose()
-    endtime = datetime.datetime.now()
-    print('All finish')
-    print('从',starttime,'开始，到',endtime,'结束，耗时为',endtime-starttime,'。共导入数据',Chinadaily.shape[0],'条。')
+    # 获取一下数据库里面最大的daily日期
+    result = engine.execute("select max(trade_date) from stock_china_daily ")
+    for row in result:
+        startdate =row['max(trade_date)']+datetime.timedelta(days=1)
+    startdate=startdate.strftime('%Y%m%d')
+    time_temp =datetime.datetime.now()
+    enddate = time_temp.strftime('%Y%m%d')
+    if startdate == enddate:
+        print('截止到%s的数据已经获取完整'%(enddate))
+    else:
+        Chinadaily = Getdailyfromtscode('',startdate,enddate)
+        Chinadaily['trade_date'] = pd.to_datetime(Chinadaily['trade_date'],format='%Y%m%d')
+        try:
+            Chinadaily.to_sql('stock_china_daily_temp', con=engine, if_exists='replace', index=False)
+        except Exception as e:
+            print(e)
+        print('数据已经导入临时表，共导入数据%s条' % (str(Chinadaily.shape[0])))
+        try:
+            query_sql = """
+                          INSERT INTO `stock_china_daily` (`ts_code`, `trade_date`, `open`, `high`, `low`, `close`, `pre_close`, `change`, `pct_chg`, `vol`, `amount`)
+                select
+                    `ts_code`, `trade_date`, `open`, `high`, `low`, `close`, `pre_close`, `change`, `pct_chg`, `vol`, `amount` from stock_china_daily_temp;
+                          """
+            query_sql2 = """
+                    delete from stock_china_daily_temp
+                """
+            engine.execute(query_sql)
+            print('数据已经导入到正是表')
+            engine.execute(query_sql2)
+            print('临时表数据已经删除')
+        except Exception as ae:
+            print(ae)
+        engine.dispose()
+        print('All finish')
+        endtime = datetime.datetime.now()
+        print('从',starttime,'开始，到',endtime,'结束，耗时为',endtime-starttime,'。共导入数据',Chinadaily.shape[0],'条。')
     '''
     db, cursor = connect_db()
     for i in range(Chinadaily.shape[0]):
